requirements.txt
fastapi
uvicorn
pydantic
pyyaml
sentence-transformers
numpy
matplotlib